Demo Numident dataset

A data set containing a sample of the CenSoc-Numident file, including age at death and select covariates.

Usage

numident_demo

Format

A data frame with 62,899 rows and 30 variables:

histid

Historical unique identifier

byear

Year of birth

bmonth

Month of birth

dyear

Year of death

dmonth

Month of death

death_age

Age at death (years)

weight

CenSoc weight

zip_residence

ZIP Code of residence at time of death

pernum

Person number in sample unit

perwt

IPUMS person weight

age

Age in 1940

sex

Sex in 1940

bpl

Place of birth

mbpl

Mother’s place of birth

fbpl

Father’s place of birth

educd

Educational attainment (detailed)

empstatd

Employment status (detailed)

hispan

Hispanic/Spanish/Latino origin

incnonwg

Had non-wage/salary income over $50

incwage

Wage and salary income

marst

Marital status

nativity

Foreign birthplace or parentage

occ

Occupation

occscore

Occupational income score

ownershp

Ownership of dwelling (tenure)

race

Race

rent

Monthly contract rent

serial

Household serial number

statefip

State of residence 1940

urban

Urban/rural status

educ_yrs

Years of education attained

Source

Joshua R. Goldstein, Monica Alexander, Casey Breen, Andrea Miranda González, Felipe Menares, Maria Osborne, Mallika Snyder, and Ugur Yildirim. CenSoc Mortality File: Version 2.0. Berkeley: University of California, 2021. https://censoc.berkeley.edu/

Steven Ruggles, Sarah Flood, Ronald Goeken, Megan Schouweiler and Matthew Sobek. IPUMS USA: Version 12.0 [dataset]. Minneapolis, MN: IPUMS, 2022. https://doi.org/10.18128/D010.V12.0

Details

The CenSoc-Numident dataset links the 1940 census to the National Archives’ public release of the Social Security Numident file. The prelinked “demo” version of the file has 63 thousand mortality records and 20 mortality covariates from the 1940 census (~1% of the complete CenSoc-Numident dataset).It is availible at https://censoc.berkeley.edu/data/.